Editing PR #101 for issue #89 resolution
The popup description should read "You have died! You have lost X coin(s)." I use the parenthesis because the player could potentially lose 1 coin. This can be omitted or it could have a check to see whether or not the `coinsLost` variable is equal to one.

It should be placed after the `coinsLost` variable. 

I would also recommend condensing the code. The line `randomCoinLoss = randomCoinLoss / 100;` is unecessary as it can simply be added directly to the `coinsLost` variable initialization.