Added logic to critical hit chance and damage
I know you originally wanted to have a 0.01% chance and 10x normal damage, but I think that 1 crit in 10 000 normal hits makes it more of an easter egg than a feature, so I've made it a 1% chance. I've further added some variability to the damage multiplier.  This of course is not my project, so if you have any objections, I will change it back to what you have intended, or make an additional if statement with your parameters.